fix:fix bugs

+++ base_curd/base_curd.py
@@ -7,7 +7,8 @@ import traceback
 from django.utils import timezone
 
 
-class BaseCURDView(RetrieveAPIView, GenericAPIView):
+class BaseCURDView(RetrieveAPIView):
+    # RetrieveAPIView 用来根据id(pk)取到对应的对象
     model = None
     serializer_class = None
     pagination_class = None
@@ -18,22 +19,23 @@ class BaseCURDView(RetrieveAPIView, GenericAPIView):
         return queryset
 
     def get_obj_by_id(self, request, *args, **kwargs):
-        # 根据id（pk）查询单个对象
+        """
+        根据id（pk）查询单个对象
+        id 是从url路径中取值，根据 lookup_field 字段
+        返回序列化好的数据
+        """
         try:
             instance = self.get_queryset().get(**kwargs)
         except self.model.DoesNotExist:
             raise Http404("对象不存在")
-        serializer = self.serializer_class(instance)
-        return serializer.data
+        ser = self.serializer_class(instance)
+        return ser.data
 
     def get_obj_list_by_conditions_page(
             self, request, exact: bool, *args, **kwargs):
         """
         使用条件查询对象列表
-        :param request:
         :param exact: 是否开启精确查询 True False
-        :param args:
-        :param kwargs:
         :return:
         """
         filter_conditions = request.query_params.dict()
@@ -74,7 +76,9 @@ class BaseCURDView(RetrieveAPIView, GenericAPIView):
             # 获取传入的ids数据
             ids_data = request.data.get("ids", [])
             if not isinstance(ids_data, list):
-                raise BaseCustomException(detail="ids必须是一个列表")
+                raise BaseCustomException("入参ids必须是一个列表")
+            if len(ids_data) == 0:
+                raise BaseCustomException("入参ids为空")
             queryset = self.get_queryset().filter(id__in=ids_data)
             queryset.delete()
             return True
@@ -87,7 +91,7 @@ class BaseCURDView(RetrieveAPIView, GenericAPIView):
             # 获取传入的ids数据
             ids_data = request.data.get("ids", [])
             if not isinstance(ids_data, list):
-                raise BaseCustomException(detail="ids必须是一个列表")
+                raise BaseCustomException(detail="入参ids必须是一个列表")
             if len(ids_data) == 0:
                 raise BaseCustomException(detail="需要删除的数据为空")
             queryset = self.get_queryset().filter(id__in=ids_data)
@@ -98,6 +102,7 @@ class BaseCURDView(RetrieveAPIView, GenericAPIView):
             raise e
 
     def update_obj_by_id(self, request, *args, **kwargs):
+        # id 是从url路径中取值，根据 lookup_field 字段
         try:
             instance = self.get_object()  # 获取要更新的对象
             data = request.data  # 获取请求的数据
